+++ test/src/analyzers/lint_analyzer/rules/rules_list/format_comment/examples/example.dart
@@ -25,3 +25,5 @@ class Test {
 //ignore
 /* with start space
 with dot.*/
+//Test multiline comment
+//second row