+++ test/src/analyzers/lint_analyzer/rules/rules_list/format_comment/format_comment_test.dart
@@ -26,8 +26,8 @@ void main() {
 
       RuleTestHelper.verifyIssues(
         issues: issues,
-        startLines: [1, 2, 3, 4, 6, 8, 11, 13, 24, 25, 26],
-        startColumns: [1, 1, 1, 1, 3, 5, 3, 5, 1, 1, 1],
+        startLines: [1, 2, 3, 4, 6, 8, 11, 13, 24, 25, 26,28,29],
+        startColumns: [1, 1, 1, 1, 3, 5, 3, 5, 1, 1, 1,1,1],
         locationTexts: [
           '// With start space without dot',
           '/* With start space without dot*/',
@@ -41,8 +41,12 @@ void main() {
           '//ignore',
           '/* with start space\n'
               'with dot.*/',
+          '//Test multiline comment',
+          '//second row',
         ],
         messages: [
+          'Prefer formatting comments like sentences.',
+          'Prefer formatting comments like sentences.',
           'Prefer formatting comments like sentences.',
           'Prefer formatting comments like sentences.',
           'Prefer formatting comments like sentences.',
@@ -68,6 +72,8 @@ void main() {
           '// Ignore.',
           '/* With start space\n'
               'with dot.*/',
+          '// Test multiline comment.',
+          '// Second row.',
         ],
       );
     });